% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_diagnostic.R
\name{sampling_diagnostic}
\alias{sampling_diagnostic}
\title{sampling_diagnostic}
\usage{
sampling_diagnostic(x, s, w = 5, pal = pal1)
}
\arguments{
\item{x}{a lazySDM object}

\item{s}{a sampling effort SpatRaster object (see example)}

\item{w}{spatial window to evaluate local effects (default=5)}

\item{pal}{a color palette (see example)}
}
\value{
A four panel plot, showing: a) global sampling effort, b)background modeled areas, c) habitability, and c) local correlation between the sampling effort and habitability.
A SpatRaster object of local correlations is also provided.
}
\description{
Diagnostic plots to evaluate the potential impact of sampling effort on SDM predictions
}
\examples{

data(whiteshark)
cmip1=terra::rast(system.file("extdata/SSP1_2.6_1985_2014.tif", package="simpleSDM"))
cmip2=terra::rast(system.file("extdata/SSP1_2.6_2070_2099.tif", package="simpleSDM"))
out=lazySDM(whiteshark,buff=500,cmip1,cmip2)

# Load a global raster of the sampling effort, i.e., number of total ocurrences in the OBIS database at one-degree resolution
# Data re-analyzed from Chaudhary et al. 2021 (https://doi.org/10.1073/pnas.2015094118)
obis=terra::rast(system.file("extdata/OBIS_sampling_effort.tif", package="simpleSDM"))
pal1=colorRampPalette(c("blue", "yellow", "red")) # a simple color palette
effortbias=sampling_diagnostic(x=out,s=log10(obis+1),w=5,pal=pal1)

# An histogram showing the distribution of local correlation values
hist(effortbias)
global(effortbias,"mean",na.rm=T)

}
