% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_white.R
\docType{data}
\name{whiteshark}
\alias{whiteshark}
\title{White Shark Occurrences}
\format{
data.frame with thre columns
\describe{
\item{x}{Mixed layer depth}
\item{y}{Primary productivity at surface}
}
}
\source{
OBIS
}
\usage{
whiteshark
}
\description{
Global georeferenced occurrences for the great white shark
}
\examples{
data(whiteshark)

}
\keyword{datasets}
