% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_pca.R
\name{stack_pca}
\alias{stack_pca}
\title{stack_pca}
\usage{
stack_pca(env1, env2, expvar = 1)
}
\arguments{
\item{env1}{Environmental stack for time 1}

\item{env2}{Environmental stack for time 2}

\item{expvar}{Proportion of variance of PCA used to select PCA axes. Default to 1 (all PCA axes)}
}
\value{
three objects
summary: PCA summary
pca.time1: PCA for environmental stack at time 1
pca.time2: PCA for environmental stack at time 2
}
\description{
converts two environmental stacks to PCA values
}
\examples{

cmip1=terra::rast(system.file("extdata/SSP1_2.6_1985_2014.tif", package="simpleSDM"))
cmip2=terra::rast(system.file("extdata/SSP1_2.6_2070_2099.tif", package="simpleSDM"))

env.pca=stack_pca(stack1,stack2,expvar=0.95)
plot(env.pca$rasters1)

}
